🐛 Retrieve replies
The script wasn't including replies in the export. (These are "thread"s
in the Gitter API)

Fix this by fetching threads if necessary. As there is no (obvious) way
to know if a thread has gotten more replies since the last export, fetch
them uncached. (This could be improved in the future by only fetching
uncached if the thread was relatively new.)